+++ db/migrate/20260621000001_add_rewritten_title_to_rewrites.rb
@@ -1,0 +1,5 @@
+class AddRewrittenTitleToRewrites < ActiveRecord::Migration[8.0]
+  def change
+    add_column :rewrites, :rewritten_title, :string
+  end
+end